requirements.txt
blinker==1.9.0
click==8.3.1
croniter==6.2.2
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.3
PyMySQL==1.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
rq==2.7.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
Werkzeug==3.1.6
